Move 'unsaved changes' window alert to shared hook code
