+++ app/webpacker/lib/hooks/useUnsavedChangesAlert.js
@@ -1,0 +1,24 @@
+import { useCallback, useEffect } from 'react';
+
+const useUnsavedChangesAlert = (hasUnsavedChanges) => {
+  const onUnload = useCallback((e) => {
+    // Prompt the user before letting them navigate away from this page with unsaved changes.
+    if (hasUnsavedChanges) {
+      const confirmationMessage = 'You have unsaved changes, are you sure you want to leave?';
+      e.returnValue = confirmationMessage;
+      return confirmationMessage;
+    }
+
+    return null;
+  }, [hasUnsavedChanges]);
+
+  return useEffect(() => {
+    window.addEventListener('beforeunload', onUnload);
+
+    return () => {
+      window.removeEventListener('beforeunload', onUnload);
+    };
+  }, [onUnload]);
+};
+
+export default useUnsavedChangesAlert;
